+++ novae/__init__.py
@@ -10,7 +10,8 @@ from . import utils
 from . import data
 from . import monitor
 from . import plot
-from .utils import spatial_neighbors, batch_effect_correction, label_domains
+from .utils import spatial_neighbors, batch_effect_correction
+from .label import label_domains
 from .data import load_dataset, toy_dataset, quantile_scaling, compute_histo_embeddings, compute_histo_pca
 
 log = logging.getLogger("novae")
