+++ novae/label/__init__.py
@@ -1,0 +1,1 @@
+from .label import label_domains
